+++ src/Rocket.java
@@ -18,7 +18,7 @@ public class Rocket extends Sprite {
      * @param img Image.
      * @param health Health.
      */
-    public Rocket(int posX, int posY, int height, int width, int velocity, Image img, int health) {
+    public Rocket(double posX, double posY, int height, int width, double velocity, Image img, int health) {
         super(posX, posY, height, width, velocity);
         this.img = img;
         this.health = health;
@@ -32,7 +32,7 @@ public class Rocket extends Sprite {
      */
     public Shot shoot() {
         int bulletVelocity = 2;
-        int xPos = posX + height / 2 - Shot.getSize() / 2;
+        double xPos = posX + height / 2 - Shot.getSize() / 2;
         int yPos = (int) posY - Shot.getSize(); // Needs to be converted from double to int
         return new Shot(xPos, yPos, height, width, bulletVelocity);
     }
